+++ opensc-opendnie/trunk/src/tools/util.h
@@ -24,7 +24,7 @@ extern "C" {
 #endif
 
 void util_print_binary(FILE *f, const u8 *buf, int count);
-void util_hex_dump(FILE *f, const u8 *in, int len, const char *sep);
+void util_hex_dump(FILE *f, const u8 *in, size_t len, const char *sep);
 void util_hex_dump_asc(FILE *f, const u8 *in, size_t count, int addr);
 void util_print_usage_and_die(const char *app_name, const struct option options[],
 	const char *option_help[]);
